data-ai-summarization/pyproject.toml
[tool.poetry]
name = "dailytide"
version = "0.1.0"
description = ""
authors = ["Kickstart.ai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
feedparser = "^6.0.11"
requests = "^2.31.0"
newsapi-python = "^0.2.7"
beautifulsoup4 = "^4.12.3"
openai = "^1.13.3"
aiohttp = "^3.9.3"
pandas = "^2.2.1"
pymongo = "^4.6.2"
tiktoken = "^0.6.0"
tqdm = "^4.66.2"
python-dotenv = "^1.0.1"
pandas-stubs = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
mypy = "^1.8.0"
black = "^24.2.0"
pre-commit = "^3.6.2"
types-requests = "^2.31.0.20240218"
types-python-dateutil = "^2.8.19.20240106"
types-tqdm = "^4.66.0.20240106"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
